+++ src/app/book-club/[bookClubSlug]/admin/members/page.tsx
@@ -1,4 +1,8 @@
-import BookClubAdminMembersTableWrapper from '@/components/wrappers/book-club-admin-members-table.wrapper';
+import { Suspense } from 'react';
+import { Spinner } from '@nextui-org/spinner';
+
+import BookClubAdminMembersTable from '@/components/tables/book-club-admin-members.table';
+import { getMembersBySlug } from '@/api/fetchers/book-club.fetchers';
 
 // Page props
 interface BookClubAdminMembersPageProps {
@@ -6,6 +10,20 @@ interface BookClubAdminMembersPageProps {
     bookClubSlug: string;
   };
 }
+
+/**
+ * Async component for rendering the admin member's table after fetching data
+ *
+ * @prop {Object} props - The component props
+ * @prop {string} props.bookClubSlug The slug of the book club
+ */
+const BookClubAdminMembersTableWrapper = async ({
+  bookClubSlug
+}: Readonly<{ bookClubSlug: string }>) => {
+  const members = await getMembersBySlug(bookClubSlug);
+
+  return <BookClubAdminMembersTable members={members} />;
+};
 
 /**
  * Page for managing members of a book club
@@ -16,8 +34,16 @@ interface BookClubAdminMembersPageProps {
  */
 const BookClubAdminMembersPage = ({
   params: { bookClubSlug }
-}: Readonly<BookClubAdminMembersPageProps>) => {
-  return <BookClubAdminMembersTableWrapper bookClubSlug={bookClubSlug} />;
-};
+}: Readonly<BookClubAdminMembersPageProps>) => (
+  <Suspense
+    fallback={
+      <div className="flex justify-center items-center w-full h-36">
+        <Spinner />
+      </div>
+    }
+  >
+    <BookClubAdminMembersTableWrapper bookClubSlug={bookClubSlug} />
+  </Suspense>
+);
 
 export default BookClubAdminMembersPage;
